fix: restore setTextureFormatToUse on WebGPU/Native engines
Move the setTextureFormatToUse / setCompressedTextureExclusions / texturesSupported /
textureFormatInUse augmentations from Engine (WebGL-only) to AbstractEngine so they
are available on every engine. Since PR #14931 (2024) WebGPUEngine no longer extends
Engine, which silently broke these calls on WebGPU.

The existing entry-point file engine.textureSelector.ts is kept as a thin
re-export for back-compat with existing side-effect imports.
