+++ packages/dev/core/src/Engines/AbstractEngine/index.ts
@@ -12,3 +12,7 @@ export * from "./abstractEngine.alpha";
 export * from "./abstractEngine.views";
 export * from "./abstractEngine.loadFile";
 export * from "./abstractEngine.textureLoaders";
+// must import first since nothing references the exports
+import "./abstractEngine.textureSelector";
+// eslint-disable-next-line no-duplicate-imports
+export * from "./abstractEngine.textureSelector";
